Move new function to util.py
